requirement.txt
python-docx==0.8.10
nltk==3.4
pattern3==3.0.0
Jinja2==2.10.1
gensim==3.7.2
spacy==2.1.3
paramiko==2.4.2
sklearn==0.0
matplotlib==3.0.3
pyLDAvis==2.1.2
Ipython